refactor(app): extract ViewState into app/view.rs
SidebarRow, PaneFocus, and ViewState with its row/selection/scroll
helpers move verbatim. search_pre_scroll and the private
next/prev_entry_index helpers become pub(super) - they are touched
by App's key handlers in mod.rs, and pub(super) preserves the
old module-private reachability exactly. The pure clamp test moves
along; the handle_key-driven tests stay in mod.rs. Part of #270.
